Swapped views -> init for deployments
