+++ ftplugin/k8s_pods.lua
@@ -1,6 +1,6 @@
 -- k8s_pods.lua in ~/.config/nvim/ftplugin
 local api = vim.api
-local deployment_view = require("kubectl.views.deployments.views")
+local deployment_view = require("kubectl.views.deployments")
 local hl = require("kubectl.actions.highlight")
 local pod_view = require("kubectl.views.pods")
 local string_util = require("kubectl.utils.string")
