+++ lua/kubectl/views/deployments/init.lua
@@ -1,41 +1,27 @@
+local definition = require("kubectl.views.deployments.definition")
+local ResourceBuilder = require("kubectl.resourcebuilder")
+
 local M = {}
-local time = require("kubectl.utils.time")
 
-function M.processRow(rows)
-  local data = {}
-  for _, row in pairs(rows.items) do
-    local pod = {
-      namespace = row.metadata.namespace,
-      name = row.metadata.name,
-      ready = M.getReady(row),
-      uptodate = row.status.updatedReplicas,
-      available = row.status.availableReplicas,
-      age = time.since(row.metadata.creationTimestamp),
-    }
-
-    table.insert(data, pod)
-  end
-  return data
+function M.Deployments()
+  ResourceBuilder:new("deployments", { "get", "deployments", "-A", "-o=json" })
+    :fetch()
+    :decodeJson()
+    :process(definition.processRow)
+    :prettyPrint(definition.getHeaders)
+    :addHints({
+      { key = "<d>", desc = "desc" },
+      { key = "<enter>", desc = "pods" },
+    }, true, true)
+    :setFilter(FILTER)
+    :display("k8s_deployments")
 end
 
-function M.getHeaders()
-  local headers = {
-    "NAMESPACE",
-    "NAME",
-    "READY",
-    "UPTODATE",
-    "AVAILABLE",
-    "AGE",
-  }
-
-  return headers
-end
-
-function M.getReady(row)
-  if row.status.availableReplicas then
-    return row.status.readyReplicas .. "/" .. row.status.availableReplicas
-  end
-  return "nil"
+function M.DeploymentDesc(deployment_desc, namespace)
+  ResourceBuilder:new("desc", { "describe", "deployment", deployment_desc, "-n", namespace })
+    :fetch()
+    :splitData()
+    :displayFloat("k8s_deployment_desc", deployment_desc, "yaml")
 end
 
 return M
